requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncclick==8.0.3.2
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.2
black==22.3.0
botometer==1.6.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
databases==0.5.5
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-decouple==3.6
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
tomli==2.0.1
tweepy==3.10.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
vaderSentiment==3.3.2
zipp==3.8.0
